+++ modules/gatsby-theme-ocular/CHANGELOG.md
@@ -1,3 +1,11 @@
+# CHANGELOG (gatsby-theme-ocular)
+
+## v1.0.1
+- Fix npmjs.org READMEs
+
+## v1.0.0
+- First publish of ocular-gatsby repackaged as a theme
+
 # CHANGELOG (ocular-gatsby)
 
 ## v1.0.0-alpha.28